Move setup() method into separate file
Add a new() function for BlockDev, to facilitate the move.
